# openapi.yaml
openapi: 3.0.3
info:
  title: Hypernova Banking API
  version: 1.0.0
  license:
    name: Proprietary technical assessment implementation
    url: https://github.com/0xStyx507/hypernova-banking
  description: |
    Contrato HTTP de Hypernova Banking.

    Incluye identidad, sesiones, auditoría, cuentas USD y movimientos
    financieros respaldados por TigerBeetle. Los importes son unidades menores
    enteras serializadas como cadenas para evitar redondeos de punto flotante.
servers:
  - url: http://127.0.0.1:8080
    description: Desarrollo local
# Public health and authentication operations opt in to security below; an
# empty root requirement makes the security posture explicit for every path.
security: []
...
paths:
...
  /api/v1/accounts:
    post:
      operationId: createAccount
      summary: Crea otra cuenta USD para el usuario autenticado
      description: Cada cuenta pertenece al usuario autenticado y puede seleccionarse como origen de una operación. La identidad recibe una cuenta inicial durante el registro.
      tags: [Ledger]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAccountRequest'
      responses:
        '201':
          description: Cuenta provisionada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '400':
          description: Solicitud inválida
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
        '401':
          description: Token inválido o ausente
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
    get:
      operationId: listAccounts
      summary: Lista las cuentas propias
      tags: [Ledger]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Cuentas del usuario
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/Account' }
        '401':
          description: Token inválido o ausente
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
...
  /api/v1/mcp/tools:
    get:
      operationId: listMCPTools
      summary: Lista herramientas disponibles para el cliente asistido
      description: Las herramientas de lectura son seguras para consulta; las mutaciones siempre requieren una acción preparada y confirmación explícita.
      tags: [MCP]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Herramientas disponibles
          content: { application/json: { schema: { $ref: '#/components/schemas/MCPToolsResponse' } } }
        '401':
          description: Token inválido o ausente
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
  /api/v1/mcp/tools/call:
    post:
      operationId: callMCPTool
      summary: Ejecuta una consulta MCP o prepara una intención
      description: Las consultas no mutan el ledger; prepare_financial_action solo crea una acción pendiente.
      tags: [MCP]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/MCPToolCallRequest' } } }
      responses:
        '200':
          description: Resultado de la herramienta
          content: { application/json: { schema: { $ref: '#/components/schemas/MCPToolCallResponse' } } }
        '400':
          description: Herramienta o argumentos inválidos
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
        '404':
          description: Herramienta inexistente
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
  /api/v1/mcp/actions:
    post:
      operationId: prepareMCPAction
      summary: Prepara una acción financiera sin ejecutarla
      description: Persiste una intención por cinco minutos. No cambia el saldo ni crea una transferencia en TigerBeetle.
      tags: [MCP]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/MCPActionRequest' } } }
      responses:
        '201':
          description: Acción lista para confirmación
          content: { application/json: { schema: { $ref: '#/components/schemas/MCPAction' } } }
        '400':
          description: Acción inválida
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
  /api/v1/mcp/actions/pending:
    get:
      operationId: getPendingMCPAction
      summary: Recupera la acción MCP pendiente del usuario
      description: Devuelve la acción lista o en confirmación después de recargar la aplicación; si no existe devuelve action null.
      tags: [MCP]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Acción pendiente o null
          content:
            application/json:
              schema:
                type: object
                required: [action]
                properties:
                  action:
                    type: object
                    nullable: true
                    allOf:
                      - $ref: '#/components/schemas/MCPAction'
        '401':
          description: Token inválido o ausente
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
  /api/v1/mcp/actions/{action_id}:
    get:
      operationId: getMCPAction
      summary: Consulta el estado de una acción preparada
      tags: [MCP]
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/ActionID'
      responses:
        '200':
          description: Acción preparada
          content: { application/json: { schema: { $ref: '#/components/schemas/MCPAction' } } }
        '404':
          description: Acción inexistente o ajena
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
  /api/v1/mcp/actions/{action_id}/confirm:
    post:
      operationId: confirmMCPAction
      summary: Confirma y ejecuta una acción financiera
      description: 'Requiere el cuerpo exacto { pin: "1234" } con un PIN previamente configurado. El ID de la acción se reutiliza como clave idempotente.'
      tags: [MCP]
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/ActionID'
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/MCPConfirmationRequest' } } }
      responses:
        '200':
          description: Acción ejecutada
          content: { application/json: { schema: { $ref: '#/components/schemas/MCPAction' } } }
        '400':
          description: Falta confirmación explícita
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
        '409':
          description: Acción no puede confirmarse en su estado actual
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
        '401':
          description: Token inválido o ausente
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
        '403':
          description: PIN MCP inválido o MFA no verificado
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
        '410':
          description: Acción preparada expirada
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
        '429':
          description: PIN MCP bloqueado temporalmente
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
        '503':
          description: Almacén de PIN o ledger no disponible
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
  /api/v1/mcp/actions/{action_id}/cancel:
    post:
      operationId: cancelMCPAction
      summary: Cancela una acción antes de su ejecución
      tags: [MCP]
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/ActionID'
      responses:
        '200':
          description: Acción cancelada
          content: { application/json: { schema: { $ref: '#/components/schemas/MCPAction' } } }
        '409':
          description: Acción ya confirmada o en estado incompatible
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
  parameters:
...
    ActionID:
      name: action_id
      in: path
      required: true
      schema: { type: string, format: uuid }
  schemas:
...
    ErrorResponse:
      type: object
      required: [error, code]
      properties:
        error:
          type: string
          description: Mensaje público y no sensible del error
          example: invalid credentials
        code:
          type: string
          description: Código estable para clientes
          example: insufficient_funds
...
    CreateAccountRequest:
      type: object
      additionalProperties: false
      properties:
        currency:
          type: string
          enum: [USD]
          default: USD
...
    Account:
      type: object
      required: [id, display_name, currency, type, status, created_at]
      properties:
        id: { type: string, format: uuid }
        display_name: { type: string, maxLength: 48 }
        currency: { type: string, enum: [USD] }
        type: { type: string, enum: [checking] }
        status: { type: string, enum: [provisioning, active, failed, closed] }
        created_at: { type: string, format: date-time }
...
    Operation:
      type: object
      required: [id, type, status, transfer_id, amount, currency, created_at]
      properties:
        id: { type: string, format: uuid }
        type: { type: string, enum: [deposit, withdrawal, transfer] }
        status: { type: string, enum: [succeeded] }
        transfer_id: { type: string }
        amount: { type: string, pattern: '^[1-9][0-9]*$' }
        currency: { type: string, enum: [USD] }
        created_at: { type: string, format: date-time }
...
    MCPTool:
      type: object
      required: [name, read_only, description]
      properties:
        name: { type: string }
        read_only: { type: boolean }
        description: { type: string }
    MCPToolsResponse:
      type: object
      required: [protocol, tools]
      properties:
        protocol: { type: string, example: hypernova-mcp-http/1 }
        tools:
          type: array
          items: { $ref: '#/components/schemas/MCPTool' }
    MCPActionRequest:
      type: object
      additionalProperties: false
      required: [action, amount]
      properties:
        action: { type: string, enum: [deposit, withdrawal, transfer] }
        account_id: { type: string, format: uuid }
        source_account_id: { type: string, format: uuid }
        destination_account_id: { type: string, format: uuid }
        transfer_type:
          type: string
          enum: [own, external]
          default: own
          description: Alcance explícito de la transferencia; el valor omitido se interpreta como own.
        amount: { type: string, pattern: '^[1-9][0-9]*$' }
        currency: { type: string, enum: [USD], default: USD }
        reason: { type: string, maxLength: 280 }
    MCPConfirmationRequest:
      type: object
      additionalProperties: false
      required: [pin]
      properties:
        pin: { type: string, pattern: '^[0-9]{4}$', description: PIN MCP de exactamente cuatro dígitos ASCII }
...
    MCPToolCallRequest:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        name: { type: string, example: get_accounts }
        arguments:
          type: object
          description: Para get_transactions se aceptan account_id, limit (1..100) y cursor. Para get_balance solo account_id.
          properties:
            account_id: { type: string, format: uuid }
            limit: { type: integer, minimum: 1, maximum: 100, default: 50 }
            cursor: { type: string, description: Cursor opaco devuelto por la página anterior }
          additionalProperties: false
    MCPToolCallResponse:
      type: object
      required: [name, result]
      properties:
        name: { type: string }
        result: {}
...
    MCPAction:
      type: object
      required: [id, action, status, payload, expires_at, created_at]
      properties:
        id: { type: string, format: uuid }
        action: { type: string, enum: [deposit, withdrawal, transfer] }
        status: { type: string, enum: [ready, confirming, confirmed, cancelled, expired, failed] }
        payload: { $ref: '#/components/schemas/MCPActionRequest' }
        expires_at: { type: string, format: date-time }
        created_at: { type: string, format: date-time }
        confirmed_at: { type: string, format: date-time }
        operation: { $ref: '#/components/schemas/Operation' }